Fix type that millis() returns

+++ Arduino15/packages/SPRESENSE/hardware/spresense/1.0.0/cores/spresense/Stream.h
@@ -51,7 +51,7 @@ class Stream : public Print
 {
   protected:
     unsigned long _timeout;      // number of milliseconds to wait for the next char before aborting timed read
-    unsigned long _startMillis;  // used for timeout measurement
+    unsigned long long _startMillis;  // used for timeout measurement
     int timedRead();    // private method to read stream with timeout
     int timedPeek();    // private method to peek stream with timeout
     int peekNextDigit(LookaheadMode lookahead, bool detectDecimal); // returns the next numeric digit in the stream or -1 if timeout
